Version 5.12.30 Beta (30.12.2013)
Version 5.12.30 Beta (30.12.2013)
Changes in the structure of the program
Database, you must re-create the program itself says on incompatibility
DB
Changed sorting algorithm DB HWID
To accelerate the launch changed selection module drivers for this
program for indexing , creates an additional file % driverpackName%.
Hwid
Clean database drivers - Removed possible errors in the analysis of inf-
files
The algorithm for the drivers on HWID in the DB
As a result of the above changes, downloading programs takes ~ 2 times
faster
Devcon_c.cmd file is no longer used when the program starts , now comes
with a direct challenge devcon.exe parameters.
Fixed a bug closing the program if the file was not created hwids.txt -
result of the program devcon.exe. The program continued to boot and gave
out a lot of errors
Fixed a bug run on windows 2000 , the reason was the function to run
external applications

Changes in the GUI
Supports Unicode for almost all controls
The default font for tips Lucida Console, for all other elements Tahoma.
Unless otherwise prescribed in the program settings .
When you switch to indexing , if not neproindeksirovannyh packages ,
this selection is set at " All "
Fixed several minor bugs - not create the folder drivers - etc
Removed all the options database drivers are now creating an
unconditional
Fixed the counting time incorrectly calculated the time in milliseconds
Removed dependency on the component richtx32.ocx - Microsoft Rich
Textbox Control 6.0
Added rebuild buttons when the window is resized or navigate through the
tabs
Changed the "About "
Added Arabic thanks MFM Dawdeh - http://www.mfmd.net/forum.php
Information about the current progress of implementation is now at the
beginning of the program header

Версия 5.12.30 Beta (30.12.2013)
Изменения в структуре программы
Базу данных необходимо пересоздать, программа сама скажет о
несовместимости БД
Изменен алгоритм сортировки БД по HWID
Для ускорения запуска изменен модуль подбора драйверов, для этого в
программе при индексации, создается дополнительный файл
%driverpackName%.hwid
Чистота базы драйверов - убраны возможные ошибки в анализе inf-файлов
Изменен алгоритм поиска драйверов по HWID в БД
В результате вышеуказанных изменений, загрузка программы происходит ~ 2
раза быстрее
Файл devcon_c.cmd более не используется при запуске программы, теперь
происходит прямой вызов devcon.exe с параметрами.
Исправлена ошибка закрытия программы, если не создавался файл hwids.txt
- результат работы программы devcon.exe. Программа продолжала
загружаться и выдавала множество ошибок
Исправлена ошибка запуска на windows 2000, причина была в функции
запуска внешних приложений

Изменения в GUI
Реализована поддержка Юникода почти для всех элементы управления
Шрифт по умолчанию для подсказок Lucida Console, для всех остальны�
элементов Tahoma. Если иного не прописано в настройках программы.
При переключении в режим индексации, если нет непроиндексированны�
пакетов, то режим выделения устанавливается во "Все"
Исправлено несколько мелких ошибок – не создание папки drivers - и т.д
Убраны все параметры создания базы драйверов, теперь создание
безусловное
Исправлена функция подсчета времени, некорректно рассчитывалось время в
миллисекунда�
Убрана зависимость от компонента richtx32.ocx - Microsoft Rich Textbox
Control 6.0
Добавлено перестроение кнопок при изменении размера окна или переходе по
вкладкам
Изменено окно "О программе"
Добавлен Арабский язык спасибо MFM Dawdeh -
http://www.mfmd.net/forum.php
Информация о текущем прогрессе выполнения теперь в начале заголовка
программы